+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(FreeRTOS)
 add_subdirectory(app)
 add_subdirectory(Init)
-add_subdirectory(HAL)
+add_subdirectory(Hal)
